Import existing formula from oms-salt-{core,tcf}
With the exception of the salt formula, the source in this import
is licensed according to:

https://github.com/IDCubed/oms-salt-core/tree/master/LICENSE
https://github.com/IDCubed/oms-salt-tcf/tree/master/LICENSE

In some cases, minor edits have been made to these formula (from what
is in the OMS), these changes may go upstream, and others may be made.
Changes apply the existing license.

This import includes:
* the python, packages, reclass, and salt formula from oms-salt-core
* the supervisord formula from oms-salt-tcf

The salt formula originally came from
https://github.com/saltstack-formulas/salt-formula