Install python-dev through the package formula generator
This includes the use of Reclass' interpolation to create a map of
package names (which sometimes change with each OS).

We include a package map for ubuntu for now.

To be clear, with this commit, we no longer need the python/dev.sls
from oms-salt-core. The rest of the python formula is needed by the
package generator formula, so we need to leave it as is.